+++ docs/CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
-
+### Fixed
+* Avoid extra re-renders on WeekView
+* Minor style fixes in WeekView
 
 
 ## 1.12.0-alpha - 2020-03-18